+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 #Changelog
 
+##2.0.1
+
+- use TLSv1 instead of SSLv3
+- add `CurlVerifier::forceSSL`
+- deprecate `CurlVerifier::forceSSLv3`
+- make calls made to `CurlVerifier::forceSSLv3` resolve to `CurlVerifier::forceSSL`
+
 ##2.0.0
 
 - Rewrite from the ground up